requirements.txt
# ============================================================
# BOTZI – Python Dependencies
# ============================================================
# Install with: pip install -r requirements.txt
# Python version: 3.11+ recommended
# ============================================================

# ── Web Framework ─────────────────────────────────────────
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pydantic==2.8.2
python-dotenv==1.0.1
starlette==0.38.4

# ── AI / LLM ──────────────────────────────────────────────
openai==1.46.0

# ── Vector Database ───────────────────────────────────────
# Pinecone v5+ with Serverless support
pinecone==5.0.1

# ── Document Parsing ──────────────────────────────────────
# PDF parsing
PyMuPDF==1.24.10
# PPTX parsing
python-pptx==1.0.1
# DOCX parsing
python-docx==1.1.2

# ── Supabase (analytics + feedback + sync log) ────────────
supabase==2.7.4

# ── HTTP client (for Zoho WorkDrive API calls) ────────────
requests==2.32.3
httpx==0.27.2

# ── Async support ─────────────────────────────────────────
anyio==4.4.0

# ── Utilities ─────────────────────────────────────────────
python-jose[cryptography]==3.3.0
